Refactor: Removed utils in favor of Metrics package
