+++ plugins/nf-co2footprint/src/main/nextflow/co2footprint/FileCreation/ReportFileCreator.groovy
@@ -5,7 +5,7 @@ import nextflow.co2footprint.CO2FootprintComputer
 import nextflow.co2footprint.Records.CO2EquivalencesRecord
 import nextflow.co2footprint.CO2FootprintConfig
 import nextflow.co2footprint.Records.CO2Record
-import nextflow.co2footprint.utils.Converter
+import nextflow.co2footprint.Metrics.Converter
 
 import groovy.text.GStringTemplateEngine
 import groovy.text.Template
