+++ plugins/nf-co2footprint/src/main/nextflow/co2footprint/Metrics/Converter.groovy
@@ -1,4 +1,4 @@
-package nextflow.co2footprint.utils
+package nextflow.co2footprint.Metrics
 
 import groovy.util.logging.Slf4j
