+++ plugins/nf-co2footprint/src/main/nextflow/co2footprint/Metrics/Quantity.groovy
@@ -1,4 +1,4 @@
-package nextflow.co2footprint.utils
+package nextflow.co2footprint.Metrics
 
 import java.math.RoundingMode
