+++ plugins/nf-co2footprint/src/test/nextflow/co2footprint/Metrics/QuantityTest.groovy
@@ -1,4 +1,4 @@
-package nextflow.co2footprint.utils
+package nextflow.co2footprint.Metrics
 
 import spock.lang.Specification
